test(ggshield): add validity check and reorganize tests
