+++ tests/cassettes/_ONE_LINE_AND_MULTILINE_PATCH.yaml
@@ -1,6 +1,7 @@
 interactions:
   - request:
-      body: '[{"filename": "test.txt", "document": "@@ -0,0 +1,29 @@\n+FacebookAppKeys:
+      body:
+        '[{"filename": "test.txt", "document": "@@ -0,0 +1,29 @@\n+FacebookAppKeys:
         294790898041573 / ce3f9f0362bbe5ab01dfc8ee565e4371 -----BEGIN RSA PRIVATE KEY-----\n+MIIBOgIBAAJBAIIRkYjxjE3KIZiEc8k4sWWGNsPYRNE0u0bl5oFVApPLm+uXQ/4l\n+bKO9LFtMiVPy700oMWLScwAN5OAiqVLMvHUCAwEAAQJANLr8nmEWuV6t2hAwhK5I\n+NNmBkEo4M/xFxEtl9J7LKbE2gtNrlCQiJlPP1EMhwAjDOzQcJ3lgFB28dkqH5rMW\n+TQIhANrCE7O+wlCKe0WJqQ3lYlHG91XWyGVgfExJwBDsAD9LAiEAmDY5OSsH0n2A\n+22tthkAvcN1s66lG+0DztOVJ4QLI2z8CIBPeDGwGpx8pdIicN/5LFuLWbyAcoZaT\n+bLaA/DCNPniBAiA0l//bzg+M3srIhm04xzLdR9Vb9IjPRlkvN074zdKDVwIhAKJb\n+RF3C+CMFb0wXme/ovcDeM1+3W/UmSHYUW4b3WYq4\n+-----END
         RSA PRIVATE KEY----- token: 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\n"}]'
       headers:
