+++ tests/cassettes/multiple_secrets.yaml
@@ -1,6 +1,7 @@
 interactions:
   - request:
-      body: '[{"document": "@@ -0,0 +1,2 @@\n+FacebookAppKeys :\n+String docker run
+      body:
+        '[{"document": "@@ -0,0 +1,2 @@\n+FacebookAppKeys :\n+String docker run
         --name geonetwork -d             -p 8080:8080 -e MYSQL_HOST=google.com             -e
         MYSQL_PORT=5434 -e MYSQL_USERNAME=root             -e MYSQL_PASSWORD=m42ploz2wd
         geonetwork\n", "filename": "test.txt"}]'
