+++ tests/cassettes/test_scan_file_secret_with_validity.yaml
@@ -1,0 +1,48 @@
+interactions:
+  - request:
+      body:
+        '[{"document": "diff --git a/test.txt b/test.txt\nnew file mode 100644\nindex
+        0000000..b80e3df\n--- /dev/null\n+++ b/test\n@@ -0,0 +2 @@\n+Sendgrid:\n+sg_key
+        = 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;\n\n",
+        "filename": "/tmp/tmpkdbj_yes/file_secret"}]'
+      headers:
+        Accept:
+          - '*/*'
+        Accept-Encoding:
+          - gzip, deflate
+        Connection:
+          - keep-alive
+        Content-Length:
+          - '280'
+        Content-Type:
+          - application/json
+        User-Agent:
+          - pygitguardian/1.1.0 (Linux;py3.8.3) ggshield
+      method: POST
+      uri: https://api.gitguardian.com/v1/multiscan
+    response:
+      body:
+        string:
+          '[{"policy_break_count":1,"policies":["File extensions","Secrets detection","Filenames"],"policy_breaks":[{"type":"SendGrid
+          Key","policy":"Secrets detection", "validity":"valid","matches":[{"type":"apikey","match":"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","index_start":36,"index_end":104,"line_start":3,"line_end":3}]}]}]'
+      headers:
+        Allow:
+          - POST, OPTIONS
+        Connection:
+          - keep-alive
+        Content-Length:
+          - '331'
+        Content-Type:
+          - application/json
+        Date:
+          - Mon, 18 May 2020 15:38:53 GMT
+        Server:
+          - nginx
+        Vary:
+          - Cookie
+        X-Frame-Options:
+          - SAMEORIGIN
+      status:
+        code: 200
+        message: OK
+version: 1
